More updates to spectrogram function

+++ cwinpy/utils.py
@@ -2,6 +2,7 @@
 General utility functions.
 """
 
+import numpy as np
 from scipy.special import gammaln
 from math import gcd
 from functools import reduce
@@ -44,7 +45,8 @@ def gcd_array(denominators):
         raise TypeError("Must have a list or array")
 
     denoms = np.asarray(denominators).flatten()  # 1d array
-    
+    denoms = denoms.astype(np.int)
+
     if len(denoms) < 2:
         raise ValueError("Must have more than two values")
